catch errors from incorrect filter application
At the moment the js client does not fully control that the selected filter is correct. (e.g. blocks that must have children, like `and`,  has them). This should be the case in the future. But, in any case, the server must also control for those (and it actually does it right now), this is required as the server must NOT trust the client and always assert that the filter is correct.

Before this commit, we let those errors propagate through the reactives a la shiny. But, if any error reaches an observe inside a module the whole application will crash. In a perfect world modules control errors in the observes and do not let this happen. But this is not a perfect world.

We solved this by placing a generic tryCatch error surrounding the calls to the filter processing that will:
 -  downgrade the errors in to warning
 - return the unfiltered dataset
 - show an error in the client using shiny::showNotification informing that no filter is applied and the nature of the error